+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Александр Гарипов on 06.01.2023.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
