+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Александр Гарипов on 06.01.2023.
+//
+
+import Foundation
+import UIKit
+
+struct QuizStepViewModel {
+  let image: UIImage
+  let question: String
+  let questionNumber: String
+}
+
